+++ src/include/storage/storage_defs.h
@@ -19,7 +19,7 @@ namespace terrier::storage {
 // All tuples potentially visible to txns should have a non-null attribute of version vector.
 // This is not to be confused with a non-null version vector that has value nullptr (0).
 #define VERSION_POINTER_COLUMN_ID col_id_t(0)
-#define NUM_RESERVED_COLUMNS 1
+#define NUM_RESERVED_COLUMNS 1u
 /**
  * A block is a chunk of memory used for storage. It does not have any meaning
  * unless interpreted by a @see TupleAccessStrategy
