+++ test/storage/storage_util_test.cpp
@@ -8,6 +8,7 @@
 #include "common/object_pool.h"
 #include "storage/data_table.h"
 #include "storage/storage_defs.h"
+#include "util/catalog_test_util.h"
 #include "util/storage_test_util.h"
 #include "util/test_harness.h"
 
@@ -204,7 +205,7 @@ TEST_F(StorageUtilTests, BlockLayoutFromSchema) {
   for (uint32_t iteration = 0; iteration < num_iterations_; iteration++) {
     uint16_t max_columns = 1000;
     const catalog::Schema schema =
-        StorageTestUtil::RandomSchema(static_cast<uint16_t>(max_columns - NUM_RESERVED_COLUMNS), &generator_);
+        CatalogTestUtil::RandomSchema(static_cast<uint16_t>(max_columns - NUM_RESERVED_COLUMNS), &generator_);
     const auto layout_and_col_map = storage::StorageUtil::BlockLayoutFromSchema(schema);
     const storage::BlockLayout layout = layout_and_col_map.first;
     const std::unordered_map<col_oid_t, col_id_t> column_map = layout_and_col_map.second;
